+++ src/luxos/api.py
@@ -23,6 +23,8 @@ def logon_required(cmd: str, commands_list=COMMANDS) -> bool | None:
 
 
 # TODO timeouts should be float | None
+# TODO prepare for refactoring using example in
+#      tests.test_asyncops.test_bridge_execute_command
 def execute_command(
     host: str, port: int, timeout: int, cmd: str, params: list[str], verbose: bool
 ):